adicionadas rotas de new e create para bookings, adicionado link na tabela para criação da reserva, adicionado criação de booking no javascript

+++ app/controllers/bookings_controller.rb
@@ -1,6 +1,23 @@
 class BookingsController < ApplicationController
+  def new
+    booking = Booking.new
+    authorize booking
+  end
+
+  def create
+    booking_datetime = "#{params[:date]} #{params[:hour]}"
+    booking = Booking.new(user: current_user, booking_datetime: booking_datetime.to_datetime)
+    authorize booking
+    if booking.save
+      render json: { message: 'Reserva realizada.', booking: booking.id }, status: 200
+    else
+      render json: { message: 'Não foi possível realizar a reserva' }, status: 422
+    end
+  end
+
   def destroy
     booking = Booking.find(params[:id])
+    authorize booking
     booking.destroy
 
     render json: { message: 'Reserva cancelada.' }, status: 200
